fsutil: Explicit parent directory creation
Currently fsutil.Create() always creates missing parent directories with
0755 mode. While this is handy for some use cases, during extraction we
want to create every path with attributes from the tarball and having
parent directories created automatically hides the bugs where we fail to
create some directories from the tarball.

One such bug is when globs are used. In this case, the directories
before wildcards are ignored when encountered in the tarball. Instead,
they are created by os.MkdirAll() in fsutil.Create() for paths matching
the glob.

Don't fix the bugs now but make the creation of the parent directories
explicit in CreateOptions with the Dirs attribute. Later, we will drop
it from call sites where appropriate.